news_addiction/requirements.txt
asgiref==3.9.1
certifi==2025.10.5
charset-normalizer==3.4.3
defusedxml==0.7.1
Django==5.2.6
django-environ==0.12.0
django-phonenumber-field==8.1.0
djangorestframework==3.16.1
djangorestframework-xml==2.0.0
dotenv==0.9.9
flake8==7.3.0
idna==3.10
mccabe==0.7.0
mysql-connector==2.2.9
mysqlclient==2.2.7
oauthlib==3.3.1
phonenumbers==9.0.15
pillow==11.3.0
pycodestyle==2.14.0
pyflakes==3.4.0
PySide6==6.9.2
PySide6_Addons==6.9.2
PySide6_Essentials==6.9.2
python-dotenv==1.1.1
requests==2.32.5
requests-oauthlib==2.0.0
shiboken6==6.9.2
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.5.0
